python_code/api/requirements.txt
openai
python-dotenv
runpod
pinecone
pandas